@@ -4,24 +4,25 @@
   "version": "0.1.0",
   "private": true,
   "dependencies": {
-    "@testing-library/jest-dom": "^5.11.9",
-    "@testing-library/react": "^11.2.5",
-    "@testing-library/user-event": "^12.8.3",
-    "axios": "^0.21.1",
-    "bootstrap": "^4.6.0",
-    "react": "^17.0.1",
-    "react-bootstrap": "^1.5.2",
-    "react-dom": "^17.0.1",
+    "@testing-library/jest-dom": "^5.16.5",
+    "@testing-library/react": "^13.4.0",
+    "@testing-library/user-event": "^14.4.3",
+    "axios": "^1.4.0",
+    "bootstrap": "^5.2.3",
+    "@babel/core": "^7.22.5",
+    "react": "^18.2.0",
+    "react-bootstrap": "^2.7.4",
+    "react-dom": "^18.2.0",
     "react-helmet": "^6.1.0",
     "react-paypal-button-v2": "^2.6.3",
-    "react-redux": "^7.2.2",
-    "react-router-bootstrap": "^0.25.0",
-    "react-router-dom": "^5.2.0",
-    "react-scripts": "4.0.3",
-    "redux": "^4.0.5",
+    "react-redux": "^8.0.5",
+    "react-router-bootstrap": "^0.26.2",
+    "react-router-dom": "^6.12.1",
+    "react-scripts": "5.0.1",
+    "redux": "^4.2.1",
     "redux-devtools-extension": "^2.13.9",
-    "redux-thunk": "^2.3.0",
-    "web-vitals": "^1.1.1"
+    "redux-thunk": "^2.4.2",
+    "web-vitals": "^3.3.2"
   },
   "scripts": {
     "start": "react-scripts start",
