@@ -1,6 +1,6 @@
 import React from "react";
 import { Container } from "react-bootstrap";
-import { BrowserRouter as Router, Route } from "react-router-dom";
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
 import Header from "./components/Header";
 import Footer from "./components/Footer";
 import HomeScreen from "./screens/HomeScreen";
@@ -22,44 +22,43 @@ import ProductEditScreen from "./screens/ProductEditScreen";
 const App = () => {
   return (
     <Router>
-      <Header />
-      <main className="py-3">
-        <Container>
-          <Route path="/order/:id" component={OrderScreen} />
-          <Route path="/placeorder" component={PlaceOrderScreen} />
-          <Route path="/payment" component={PaymentScreen} />
-          <Route path="/shipping" component={ShippingScreen} />
-          <Route path="/profile" component={ProfileScreen} />
-          <Route path="/register" component={RegisterScreen} />
-          <Route path="/login" component={LoginScreen} />
-          <Route path="/product/:id" component={ProductScreen} />
-          <Route path="/cart/:id?" component={CartScreen} />
-          <Route path="/admin/userlist" component={UserListScreen} />
-          <Route path="/admin/user/:id/edit" component={UserEditScreen} />
+    <Header />
+    <main className="py-3">
+      <Container>
+        <Routes>
+          <Route path="/order/:id" element={<OrderScreen />} />
+          <Route path="/placeorder" element={<PlaceOrderScreen />} />
+          <Route path="/payment" element={<PaymentScreen />} />
+          <Route path="/shipping" element={<ShippingScreen />} />
+          <Route path="/profile" element={<ProfileScreen />} />
+          <Route path="/register" element={<RegisterScreen />} />
+          <Route path="/login" element={<LoginScreen />} />
+          <Route path="/product/:id" element={<ProductScreen />} />
+          <Route path="/cart/:id?" element={<CartScreen />} />
+          <Route path="/admin/userlist" element={<UserListScreen />} />
+          <Route path="/admin/user/:id/edit" element={<UserEditScreen />} />
           <Route
             path="/admin/productlist/:pageNumber"
-            component={ProductListScreen}
-            exact
+            element={<ProductListScreen />}
           />
           <Route
             path="/admin/productlist"
-            component={ProductListScreen}
-            exact
+            element={<ProductListScreen />}
           />
-          <Route path="/admin/orderlist" component={OrderListScreen} />
-          <Route path="/admin/product/:id/edit" component={ProductEditScreen} />
-          <Route path="/search/:keyword" component={HomeScreen} exact />
-          <Route path="/page/:pageNumber" component={HomeScreen} exact />
+          <Route path="/admin/orderlist" element={<OrderListScreen />} />
+          <Route path="/admin/product/:id/edit" element={<ProductEditScreen />} />
+          <Route path="/search/:keyword" element={<HomeScreen />} />
+          <Route path="/page/:pageNumber" element={<HomeScreen />} />
           <Route
             path="/search/:keyword/page/:pageNumber"
-            component={HomeScreen}
-            exact
+            element={<HomeScreen />}
           />
-          <Route path="/" component={HomeScreen} exact />
-        </Container>
-      </main>
-      <Footer />
-    </Router>
+          <Route path="/" element={<HomeScreen />} />
+        </Routes>
+      </Container>
+    </main>
+    <Footer />
+  </Router>
   );
 };
 
